Split IP address validation utils into their own file
