+++ staging/src/k8s.io/apimachinery/pkg/util/validation/validation_test.go
@@ -322,108 +322,6 @@ func TestIsValidLabelValue(t *testing.T) {
 	}
 }
 
-func TestIsValidIP(t *testing.T) {
-	goodValues := []string{
-		"::1",
-		"2a00:79e0:2:0:f1c3:e797:93c1:df80",
-		"::",
-		"2001:4860:4860::8888",
-		"::fff:1.1.1.1",
-		"1.1.1.1",
-		"1.1.1.01",
-		"255.0.0.1",
-		"1.0.0.0",
-		"0.0.0.0",
-	}
-	for _, val := range goodValues {
-		if msgs := IsValidIP(val); len(msgs) != 0 {
-			t.Errorf("expected true for %q: %v", val, msgs)
-		}
-	}
-
-	badValues := []string{
-		"[2001:db8:0:1]:80",
-		"myhost.mydomain",
-		"-1.0.0.0",
-		"[2001:db8:0:1]",
-		"a",
-	}
-	for _, val := range badValues {
-		if msgs := IsValidIP(val); len(msgs) == 0 {
-			t.Errorf("expected false for %q", val)
-		}
-	}
-}
-
-func TestIsValidIPv4Address(t *testing.T) {
-	goodValues := []string{
-		"1.1.1.1",
-		"1.1.1.01",
-		"255.0.0.1",
-		"1.0.0.0",
-		"0.0.0.0",
-	}
-	for _, val := range goodValues {
-		if msgs := IsValidIPv4Address(field.NewPath(""), val); len(msgs) != 0 {
-			t.Errorf("expected %q to be valid IPv4 address: %v", val, msgs)
-		}
-	}
-
-	badValues := []string{
-		"[2001:db8:0:1]:80",
-		"myhost.mydomain",
-		"-1.0.0.0",
-		"[2001:db8:0:1]",
-		"a",
-		"2001:4860:4860::8888",
-		"::fff:1.1.1.1",
-		"::1",
-		"2a00:79e0:2:0:f1c3:e797:93c1:df80",
-		"::",
-	}
-	for _, val := range badValues {
-		if msgs := IsValidIPv4Address(field.NewPath(""), val); len(msgs) == 0 {
-			t.Errorf("expected %q to be invalid IPv4 address", val)
-		}
-	}
-}
-
-func TestIsValidIPv6Address(t *testing.T) {
-	goodValues := []string{
-		"2001:4860:4860::8888",
-		"2a00:79e0:2:0:f1c3:e797:93c1:df80",
-		"2001:0db8:85a3:0000:0000:8a2e:0370:7334",
-		"::fff:1.1.1.1",
-		"::1",
-		"::",
-	}
-
-	for _, val := range goodValues {
-		if msgs := IsValidIPv6Address(field.NewPath(""), val); len(msgs) != 0 {
-			t.Errorf("expected %q to be valid IPv6 address: %v", val, msgs)
-		}
-	}
-
-	badValues := []string{
-		"1.1.1.1",
-		"1.1.1.01",
-		"255.0.0.1",
-		"1.0.0.0",
-		"0.0.0.0",
-		"[2001:db8:0:1]:80",
-		"myhost.mydomain",
-		"2001:0db8:85a3:0000:0000:8a2e:0370:7334:2001:0db8:85a3:0000:0000:8a2e:0370:7334",
-		"-1.0.0.0",
-		"[2001:db8:0:1]",
-		"a",
-	}
-	for _, val := range badValues {
-		if msgs := IsValidIPv6Address(field.NewPath(""), val); len(msgs) == 0 {
-			t.Errorf("expected %q to be invalid IPv6 address", val)
-		}
-	}
-}
-
 func TestIsHTTPHeaderName(t *testing.T) {
 	goodValues := []string{
 		// Common ones
@@ -706,34 +604,6 @@ func TestIsDomainPrefixedPath(t *testing.T) {
 	}
 	for _, val := range badValues {
 		if err := IsDomainPrefixedPath(field.NewPath(""), val).ToAggregate(); err == nil {
-			t.Errorf("expected errors for %q", val)
-		}
-	}
-}
-
-func TestIsValidSocketAddr(t *testing.T) {
-	goodValues := []string{
-		"0.0.0.0:10254",
-		"127.0.0.1:8888",
-		"[2001:db8:1f70::999:de8:7648:6e8]:10254",
-		"[::]:10254",
-	}
-	for _, val := range goodValues {
-		if errs := IsValidSocketAddr(val); len(errs) != 0 {
-			t.Errorf("expected no errors for %q: %v", val, errs)
-		}
-	}
-
-	badValues := []string{
-		"0.0.0.0.0:2020",
-		"0.0.0.0",
-		"6.6.6.6:909090",
-		"2001:db8:1f70::999:de8:7648:6e8:87567:102545",
-		"",
-		"*",
-	}
-	for _, val := range badValues {
-		if errs := IsValidSocketAddr(val); len(errs) == 0 {
 			t.Errorf("expected errors for %q", val)
 		}
 	}
